requirements.txt
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
folium==0.20.0
geographiclib==2.1
geopy==2.4.1
googletrans==4.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
pyzmq==27.1.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
textblob==0.19.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
xyzservices==2025.4.0
yarg==0.1.9
